Change power and location fields from integers
Latitude, longitude, and power table columns where integers before.

+++ alembic/versions/1faf6c2610bb_create_live_network_umts_cells_data_.py
@@ -34,7 +34,7 @@ def upgrade():
         sa.Column('lac', sa.Integer,),
         sa.Column('maximum_transmission_power', sa.Double,),
         sa.Column('secondary_sch_power', sa.Double, ),
-        sa.Column('azimuth', sa.Integer,),
+        sa.Column('azimuth', sa.Float,),
         sa.Column('height', sa.Integer,),
         sa.Column('primary_sch_power', sa.Integer,),
         sa.Column('rac', sa.Integer,),
